requirements.txt
certifi>=2021.5.30
charset-normalizer>=2.0.12
click>=8.0.4
cycler>=0.11.0
docker-pycreds>=0.4.0
gitdb>=4.0.9
GitPython>=3.1.18
grad-cam>=1.3.6
idna>=3.3
importlib-metadata>=4.8.3
importlib-resources>=5.4.0
joblib>=1.0.1
kiwisolver>=1.3.1
matplotlib>=3.3.4
mkl-fft>=1.3.0
mkl-random>=1.1.1
mkl-service>=2.3.0
numpy>=1.19.5
opencv-python>=4.5.5.64
pathtools>=0.1.2
Pillow>=8.4.0
pip>=21.2.2
promise>=2.3
protobuf>=3.19.4
psutil>=5.9.0
pyparsing>=3.0.9
python-dateutil>=2.8.2
pytorch-gradcam>=0.2.1
PyYAML>=6.0
requests>=2.27.1
scikit-learn>=0.24.2
scipy>=1.5.2
sentry-sdk>=1.5.9
setproctitle>=1.2.2
setuptools>=58.0.4
shortuuid>=1.0.8
six>=1.16.0
smmap>=5.0.0
tensorboardX>=2.5
threadpoolctl>=2.2.0
torch>=1.8.0+cu111
torchaudio>=0.8.0
torchvision>=0.9.0+cu111
tqdm>=4.64.0
ttach>=0.0.3
typing_extensions>=4.1.1
urllib3>=1.26.9
wandb>=0.12.14
wheel>=0.37.1
zipp>=3.6.0
flwr[simulation]>=1.8.0
flwr-datasets[vision]==0.0.2
torch==2.2.1
torchvision==0.17.1